+++ File.h
@@ -60,7 +60,7 @@ public:
   int IsEol() const;
   char *ReadLine();
   char *ReadString();
-  int ReadSizeT();
+  size_t ReadSizeT();
   int ReadInt();
   double ReadDouble();
   void ReadKeyword(const char *keyword);
@@ -110,7 +110,7 @@ public:
   char *GetName();
 
   // Write function
-  void Write(const int v, const char *sep = NULL);
+  void Write(const size_t v, const char *sep = NULL);
   void Write(const int v, const char *sep=NULL);
   void Write(const double v, const char *sep=NULL);
   void Write(const char *s);
