+++ GLApp/GLCombo.h
@@ -16,17 +16,17 @@ public:
 
   // Component methods
   void Clear();
-  void SetSize(int nbRow);
-  void SetValueAt(int row, const std::string& value, int userValue = 0);
-  int  GetUserValueAt(int row);
+  void SetSize(size_t nbRow);
+  void SetValueAt(size_t row, const std::string& value, int userValue = 0);
+  int  GetUserValueAt(size_t row);
   void SetSelectedValue(const char *value);
   void ScrollTextToEnd();
   void SetSelectedIndex(int idx);
   int  GetSelectedIndex();
   std::string GetSelectedValue();
   void SetEditable(bool editable);
-  std::string GetValueAt(int row);
-  int  GetNbRow();
+  std::string GetValueAt(size_t row);
+  size_t  GetNbRow();
 
   // Implementation
   void Paint();
